adding Dto and dao, schema sql to create necssary data and app.properties to start the db2 DB
