+++ src/main/java/org/mystore/controller/product/ProductController.java
@@ -1,6 +1,6 @@
 package org.mystore.controller.product;
 
-import org.mystore.dao.ProductDao;
+import org.mystore.dto.ProductDto;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -10,7 +10,7 @@ public class ProductController {
     private static final String PATH = "/product";
 
     @PostMapping(value = PATH)
-    public String insertProduct(ProductDao product){
+    public String insertProduct(ProductDto product){
         return "null";
     }
 }
