+++ src/main/java/org/mystore/dao/ProductDao.java
@@ -1,34 +1,21 @@
 package org.mystore.dao;
 
-import org.mystore.Exceptions.ProductDescriptionException;
+import org.mystore.dto.ProductDto;
+
 public class ProductDao {
-    public static final String INCORRECT_PRODUCT_NAMING_MESSAGE = "Incorrect product naming";
-    public static final String PRICE_CANNOT_BE_A_NEGATIVE_VALUE_MESSAGE = "Price cannot be a negative value";
-    private String name;
-    private Long princeInCents;
+    private Integer id;
+    private ProductDto productDto;
 
-    public String getName(){
-        return this.name;
+    public Integer getId() {
+        return id;
     }
 
-    public Long getPrinceInCents(){
-        return this.princeInCents;
+    public ProductDto getProductDto() {
+        return productDto;
     }
 
-    public ProductDao(String name, Long princeInCents){
-        if (name != null && !name.isBlank())
-            this.name = name;
-        else throw new ProductDescriptionException(INCORRECT_PRODUCT_NAMING_MESSAGE);
-        if (princeInCents >= 0 )
-            this.princeInCents = princeInCents;
-        else throw new ProductDescriptionException(PRICE_CANNOT_BE_A_NEGATIVE_VALUE_MESSAGE);
-    }
-
-    @Override
-    public String toString() {
-        return "Product {" +
-                "name='" + name + '\'' +
-                ", princeInCents=" + princeInCents +
-                '}';
+    public ProductDao(Integer id, String name, Long priceInCents) {
+        this.id = id;
+        this.productDto = new ProductDto(name, priceInCents);
     }
 }
